Add security to views

+++ menu/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.urls import reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import MenuItem, Category, DietaryRestriction
 from django.db.models import Q
 from collections import defaultdict
@@ -73,8 +74,13 @@ def menu_item_detail(request, menu_item_id):
     return render(request, 'menu/menu_item_detail.html', context)
 
 
+@login_required
 def add_menu_item(request):
     """ Add a menu item to the menu """
+    if not request.user.is_superuser:
+        messages.error(request, "Sorry, only admins can do that.")
+        return redirect(reverse('home'))
+    
     if request.method == 'POST':
         form = MenuItemForm(request.POST, request.FILES)
         if form.is_valid():
@@ -93,9 +99,13 @@ def add_menu_item(request):
 
     return render(request, template, context)
 
-
+@login_required
 def edit_menu_item(request, menuitem_id):
     """ Edit a menu item """
+    if not request.user.is_superuser:
+        messages.error(request, "Sorry, only admins can do that.")
+        return redirect(reverse('home'))
+
     menuitem = get_object_or_404(MenuItem, pk=menuitem_id)
     if request.method == 'POST':
         form = MenuItemForm(request.POST, request.FILES, instance=menuitem)
@@ -117,9 +127,13 @@ def edit_menu_item(request, menuitem_id):
 
     return render(request, template, context)
 
-
+@login_required
 def delete_menu_item(request, menuitem_id):
     """ Delete a menu item """
+    if not request.user.is_superuser:
+        messages.error(request, "Sorry, only admins can do that.")
+        return redirect(reverse('home'))
+
     menuitem = get_object_or_404(MenuItem, pk=menuitem_id)
     menuitem.delete()
     messages.success(request, 'Menu Item deleted!')
